Expose library code under the torchbiggraph package
Summary:
Use the external package name internally as well, thus renaming it from `deeplearning.projects.filament2` to `torchbiggraph`. This makes life easier for code that will be open-sourced but doesn't belong to the main library (tests, converters, ...), as it'll be able to load the main code from the same package name without too much "magic". It also reduces overall ambiguity and confusion.

This implies that the Buck targets now don't match the package name. On one hand, it means people just running torchbiggraph won't have to change the target they use. On the other hand, this discrepancy is a bit annoying. We may fix it one day...

Reviewed By: adamlerer

Differential Revision: D13693858

fbshipit-source-id: 86d9f0b17e81865515077cb923d715ed4cdda6a9