Construct job_tmpdir in a separate operator
Summary:
This is a follow-up to D13558348 . As a consequence of D13558348 , if you don't set any paths (modeldir etc.) it will put them all in a gluster tempdir (job_tempdir), but if you clone that run it will have to start from scratch in a new tempdir.

As a result of this diff, if you clone a run and select "Resume this run and copy all successful operators", it will re-use the same job_tmpdir and can continue successfully.

Users can also look in Outputs which will show job_tmpdir, so they can continue from this tmpdir by setting job_tmpdir=<job_tmpdir> in a future run.

Thanks to lerks for the idea.

Reviewed By: lerks

Differential Revision: D13722113

fbshipit-source-id: 3ba4bf89c025cc2236bc7a063d877a78f8b6d5c5