Have trainer just yield stats rather than pushing to queue
Summary:
Still as part of the open-sourcing effort, this diff further decouples torchbiggraph from how it's being used in the filament2 workflow. Instead of providing stats and updates in the format the workflow wants (i.e., pushed on a queue, represented as a dict), it uses a more "standard" interface (a Python generator, yielding torchbiggraph's own representations).

Also do the same for evaluation, for harmonization and because it will soon be useful for tests to easily access values that are otherwise only printed to logs.

It turns out that when you add type annotations to the signature of a function, the linter starts type checking the body of that function, which triggered a few errors that I'm fixing.

Reviewed By: adamlerer

Differential Revision: D13732310

fbshipit-source-id: 07af1bbcb222af79709b818853c7eccce9949328