initialize and load tensors directly into shared memory
Summary:
During embedding initialization and swapping, we first load the tensor then copy to shared memory. This leads to a temporary spike where we're using up to 2x memory, which leads to OOM in memory-constrained cases.

This seems to fix the problem. It's somewhat hacky in the loading case bug I don't have a better approach at the moment. lerks will probably have a more Python-y way to do it.

P.S. I haven't done this for the EdgeLoader case. It's not as important, and requires a different set of tricks (for hdf5/numpy), so I'm punting this for now.

Thanks to colesbury for the pointer to `_new_shared()`.

Reviewed By: lerks

Differential Revision: D13883547

fbshipit-source-id: 3d62f74f40a3f796cbf05404ff562d7d1c1dc0e4