Make config types as narrow as possible
Summary:
A ton of input parameters to the workflow had an `Optional` type although it made no sense for their value to be `None`. I don't know the reason for this (maybe someone though that optional meant "can be omitted and fall back to default"?) but it feels clearly wrong: in the best case one got an accidentally correct behaviors (in a boolean context `None` is falsy), in the worst case something crashed, in the medium case one ended up doing queries `SELECT ... FROM None`.

This diff removes `Optional` from wherever `None` isn't a legal value (to my knowledge).

It does it to both the workflow and the PBG code (which suffered from a similar condition), as the two need to be kept in sync. It also adds new syntactic sugar to carry over the defaults from PBG to the workflow.

Finally, it also adds type annotations to the last functions of the workflow that didn't have it already, and it uses relative imports there for brevity.

Reviewed By: adamlerer

Differential Revision: D13898828

fbshipit-source-id: 0691e6a2fa7824f3152fb326cdd14c5457915c49