Fix inability to parse configs
Summary:
As a last-minute change in D13956860 I modified the name of the function that config modules need to expose, so that it contains the `torchbiggraph` substring, making it more greppable in the future. It turns out I forgot to change the place where we called that function.

Note: this didn't affect any of the FBLearner users, as the flow produces its config all by itself.

Reviewed By: adamlerer

Differential Revision: D13988195

fbshipit-source-id: db235233fb8c205d5ac6feb911c3b40531632681