Don't always generate initial embeddings in the positive quadrant
Summary:
Unless a prior was provided, we initialized embeddings in `[0, eps)^D`. This was a bit odd: why prefer the positive quadrant? This diff samples initial embeddings from `D` "indipendent" centered normal distributions of standard deviation `eps`. (The dimensions of each entities are independent, but across different entities they may be correlated).

Also, sometimes initial embeddings are also generated when downloading the entities (if an initial value is given for some but not all entities). We need to make sure these embeddings are created the same way. This requires some changes to the workflow that I'm not too excited about. If anyone has a better idea please let me know.

Reviewed By: adamlerer

Differential Revision: D14041595

fbshipit-source-id: 86fc4199b4245c5bb6e307ec9761434435e45c5a