Make bucket generation repeatable
Summary: Contbuild caught (after the push) a flaky test due to a non-deterministic bug in the new bucket ordering code I added in D14022643. I managed to reproduce the bug and fixed it, and I also allowed that code to run determinstically by having it take a generator that can be seeded. The tests now output the seed they used when they fail, easing debugging.

Reviewed By: adamlerer

Differential Revision: D14080954

fbshipit-source-id: e0cc202694e316c5e0dda6fe277019ec91231828