Have the user manually activate dynamic relation mode
Summary: We used to always check whether there was a `dynamic_rel_count.pt` file and in that case turn on the mode, with the number of relations we found in that file. This led there to be a field in the config that users weren't supposed to set. This was awkward. Now that field has become a bool, and users are require to set it if they want PBG to activate the mode and look for the above file.

Reviewed By: adamlerer

Differential Revision: D14064701

fbshipit-source-id: 6a452219080a431581a7b22a3126578e2e67a524