In inside_out and outside_in, make transposed buckets adjacent
Summary: Affinity (i.e., number of common partitions) of successive buckets should be high as this saves the unloading and reloading of embeddings from disk. The inside_out and outside_in methods were not trying to have pairs with affinity 2 (the highest one) be next to each other. This now fixes it.

Reviewed By: adamlerer

Differential Revision: D14081723

fbshipit-source-id: eeb7de1b5dcdc6e3939ef7ae56fb69f799435580