No race condition when switching checkpoint versions
Summary:
My understanding is that in distributed mode the checkpoint could have become corrupted if some worker (except the rank-0 one) died during checkpointing, as in that case the rank-0 worker would update the version file even though not all files for that version had been written. The data of the old version wouldn't be deleted, this scenario wouldn't cause data loss, but it would require manual intervention to decrease the version by 1.

This can be easily fixed by adding one more barrier between writing all the files and updating the version.

Reviewed By: adamlerer

Differential Revision: D14149882

fbshipit-source-id: 98f28103cfc1a4c8a18537c1db038facdaed437d