Store entity partition embeddings as HDF5
Summary:
We're finally getting to the meat of this stack! This commit stops storing the embeddings of a partition of the entities (and the state dict of their associated optimizer) as a PyTorch-style picked tuple and instead used a HDF5 file. See the task for the rationale.

The embeddings are stored as a standard HDF5 dataset. This allows to "handle them" without having to load them fully in memory (see the task for what benefits this brings). The optimizer state dict is hard to map to HDF5 structures as it's an opaque piece of data to PBG's eyes (in practice, it's a nested dict with str keys and scalar or tensor values). So it is best handled by first passing it through `torch.save` to get a binary dump, then store that as a 1-dimensional byte dataset. This introduces some overhead. When writing, we must first save the data entirely in memory becuase to create a (non-chunked) dataset we must know its final size beforehand. Reading is better because by wrapping the dataset in a file-like object we can read it as we need it and keep a smaller overhead.

Moreover, we can add a whole bunch of metadata to these files which we didn't have in the tuples.

Reviewed By: adamlerer

Differential Revision: D14183426

fbshipit-source-id: b3a573efdd8f70a9abb36e6ed9d945633c683906