Store metadata as HDF5
Summary: The remaining PyTorch-style pickle to be converted to HDF5 is the metadata file. It contains the model and the optimizer state dict. The latter is stored as a 1-dim byte dataset of its pickle, but the model state dict is more useful to users so it is stored as separate datasets (the fact that the model state dict is a shallow str-to-tensor dict makes this easier). The keys of the state dict however are converted to "public" names, so that we don't expose the internal model implementation details. The code is a bit ugly as it tries to match each key of the state dict to a public key, but at least this will ensure that if a new key is added that we don't know about we won't just silently ignore it.

Reviewed By: adamlerer

Differential Revision: D14183425

fbshipit-source-id: 3eebc6d6c0047665f1b5733195a0830162f5cbfb