Remove indices of epoch, edge path and edge chunk from metadata
Summary: This information is already contained in the checkpoint version, which is just the overall index of the tuple `(epoch_idx, edge_path_idx, echunk_idx)`. A later diff in the stack will bring this information back into the metadata file, but only intended for human consumption, as PBG will continue to rely only on the checkpoint version to decide where to resume training from.

Reviewed By: adamlerer

Differential Revision: D14183427

fbshipit-source-id: 0b9bd71241d4b3c50bdb926229e8d3a4d03ec008