Introduce IterationManager to separate the training iteration logic
Summary: The training loops iterates on the product of epochs, edge paths and edge chunks, and the checkpoint version is the index in this list. This isn't magic or complex, but we move it out to a separate class as this makes it easier for a future diff in the stack to add this information as metadata to all checkpoint files.

Reviewed By: adamlerer

Differential Revision: D14183424

fbshipit-source-id: fce132d6c2fd98e67b425ee3b3a5a05ceec86a17