Store config in a JSON file separate from the metadata
Summary: The config doesn't change during the training loop, so it's not necessary to embed a copy into each metadata file. Moreover, storing it in a pickled format makes it hard for humans to look it up. Storing it in a separate JSON-encoded file makes it more useful for humans (and equally easy for machines). A later diff in the stack will bring back the config in the metadata file, but just for redundnacy.

Reviewed By: adamlerer

Differential Revision: D14183428

fbshipit-source-id: f2519b61ea1249026728fcbd14c8bf5f577d941d