Give warnings when it appears the old I/O format is being used
Summary:
**IF YOU ARRIVED HERE BY FOLLOWING THE WARNING MESSAGE**
This diff is at the top of a stack that converted the I/O format of PBG (a.k.a. Filament2). If you have data in the old format, you will need to convert it manually before being able to run new versions of PBG on it.
* You can convert the entity path by running `//deeplearning/projects/filament2:transform_entity_path`.
* You can convert the edge paths by running `//deeplearning/projects/filament2:transform_edge_path`.
* You can convert the checkpoint path and the init path by running `//deeplearning/projects/filament2:transform_checkpoint_path`.

All these take as first argument an existing directory with the old data, and as second argument a non-existing subdirectory of an existing directory where the new data will be written to. Once the conversion is complete you need to either change the config to point to the new locations, or move the data back to the old locations.

Diff description:

To help transition people over to the new format, we provided converters in D14183421. They should trigger for everyone using FBLearner, so they will mostly be transparent to users, but they won't kick in in case of manual operation. To put unaware users on the right track, I'm adding autodetection of the old format which will print warnings leading to this diff.

Reviewed By: adamlerer

Differential Revision: D14241362

fbshipit-source-id: 9bdff892094a30bdfd1e4bcbe119dfd176b3961e