Add functional tests for partitioned entities
Summary:
The title says it all.

This uncovered that there was a check preventing the number of lhs partitions from being different than the number of rhs ones. As I don't see any reason why that should be, I reworked that check so that this setting is now allowed.

I also fixed an issue with the auto-detection of the old I/O format.

Reviewed By: adamlerer

Differential Revision: D14303565

fbshipit-source-id: 699be3b67f9f7e41d36a16fa68ede834f6ef9d11