Add functional tests for featurized entities
Summary:
The title says it all.

To avoid a combinatorial explosion of tests we keep them "modular", i.e., each test focuses only on one feature, in isolation, assuming all other ones work. This is why we're not testing the interactions between featurized entities and, say, partitions, or the "all-negatives" mode, etc. This may cause us not to catch some problems, but also reduces and simplifies tests.

While writing this I temporarily introduced a bug in the generator that produced empty partitions. This uncovered a few issues with code not being able to handle empty data. I fixed the ones I hit. I also had an issue with some code in TensorList trying to modify a tensor but being unable to becuase it was a PyTorch view of a numpy ndarray and was thus "read-only". That ndarray was coming from HDF5. To avoid this, I'm now first creating a PyTorch tensor, then giving HDF5 a ndarray that is a view on the tensor's data. This way we get the HDF5 data into a regular tensor. There were also some issues with empty long tensors not having any dtype specified at creation, thus defaulting to float tensors. Although this wasn't causing issues, I fixed that.

Reviewed By: adamlerer

Differential Revision: D14303518

fbshipit-source-id: 1aaed08da19fccd8919deeaf29407fca5572fda4