Don't use specialized subclasses of Tensor in type annotations
Summary:
The type stub of PyTorch defines a Tensor class, with all its supported methods, all of which accept or return other Tensors. It then defines subclasses Tensor (LongTensor, FloatTensor, ...), which are empty. What this means is that those subclasses also have all the Tensor methods, but their annotations only specify that they return Tensors, rather than a more specific typed subclass.

The fact that we used the specialized subclasses in our annotations causes a lot of typing errors (in, say PyCharm) about "expected LongTensor, got Tensor".

To make our typing annotations "work" without making them less informative (at least for humans), this diff defines LongTensorType and FloatTensorType as aliases of Tensor. This way the tools will treat them as Tensors and will stop complaining, but the information on the type is still there (although unenforced).

PS: the LongTensor&co. classes are described as legacy (in the code and by the PyTorch devs I spoke to), with the use of Tensor only being encouraged (which is also what NumPy does), thus the PyTorch stub isn't expected to improve w.r.t. LongTensor and FloatTensor.

Reviewed By: adamlerer

Differential Revision: D14439120

fbshipit-source-id: 738fcc797ce54e2d954ee58f9c449785e1542967