Use "new" tensor creation ops and be consistent in shape and dtype args
Summary:
From my chats with PyTorch devs, it seems that constructing tensor objects using `torch.Tensor` is discouraged (perhaps even deprecated?), as it appears way too overloaded and prone to confusion (Q: does `Tensor((2, 3))` construct an empty 2x3 tensor or a 1-d tensor with two elements, 2 and 3? A: the latter. Q: of what type? A: float.).

The same holds for FloatTensor, LongTensor, etc. and even more so, as these whole classes are legacy.

The documentation doesn't even report these construction operators, suggesting instead the `tensor`, `empty`, `zeros`, `full`, etc. functions. These appears to be what is recommended nowadays.

This diff moves away from the old-style creation ops (these were plenty especially in parameterserver.py). It also consistently passes a `dtype` argument to the new-style functions (instead of relying on the fact that `float` is the default for floating-point data and `long` for integer data, except that this doesn't work for empty tensors and could break when `set_default_dtype` or `set_default_tensor_types` are used). Finally, also some of the new ops are overloaded (e.g., `ones(2, 3)` is the same as `ones((2, 3))`). This diff consistently uses a tuple, as that's the only syntax that works everywhere (because `full` doesn't accept a list of arguments).

Reviewed By: adamlerer

Differential Revision: D14439167

fbshipit-source-id: 8e5907472b17dcc7be63423c5461a5e84f23561f