Mirror PBG doc in Wiki
Summary: By just providing a Buck TARGET, we can have the full doc of PBG be automatically picked up by a cron job, built as Wiki pages and published. The results are not as good as the "real" doc (images aren't picked up, math isn't rendered, ...), but at least it makes this content findable and indexed internally.

Reviewed By: adamlerer

Differential Revision: D14538764

fbshipit-source-id: 602fe20bbfd6eaec4acc164b8096f79438d5a828