Factor out distributed mode checks for barriers
Summary:
Trainers in distributed mode synchronize through barriers. So we had code like
```
if distributed:
    trainers = ...
step1()
if distributed:
    barrier(trainers)
step2()
```
This made different linters unhappy in different ways:
- Flake8 complained about high complexity of the function, because each of these ifs introduced a branching.
- PyCharm complained that trainers wasn't defined in all branches and thus it couldn't be sure that it existed when used. (It did, because the check was the same in all ifs, but PyCharm couldn't prove that).

With a small abstraction we make everyone happy: we define two "barrier providers" with the same interface, we instantiate a dummy one if non-distributed and a real one otherwise.

Reviewed By: adamlerer

Differential Revision: D14520722

fbshipit-source-id: 7fff819b1f987d6a8f4770111df449722ea6504e