Factor code for bucket scheduling out of train loop
Summary: The train function contained code to decide what bucket to operate on that had to work very differently based on whether the system was distributed (i.e., create_ordered_buckets vs a full LockServer). It did so by repeatedly branching. Creating two classes with the same interface for the two cases simplifies the training loop a lot.

Reviewed By: adamlerer

Differential Revision: D14520741

fbshipit-source-id: da9a6d46676515c024ed8b8c95b94724464ce9d2