Some more small changes to distributed mode
Summary:
Move stuff for distributed mode setup out of util and train into dedicated module. This includes the rank assignment and the initialization code. This allows the partitionserver to also use it, and avoid duplicating the rank-assignment logic.

Also, add checks to both trainer and partition server that their rank is correct, and fail with a good error message when PyTorch doesn't come with c10d support (happens for example on Mac OS).

Reviewed By: adamlerer

Differential Revision: D14520844

fbshipit-source-id: 124e5080845da1ecd61c79fb8b360f0e15d1eae6