Validate the config
Summary:
When parsing the config perform some easy sanity checks to detect silly errors early. Type checks are already performed by the dict-to-schema loader, so we only need to check the values.

This also produces a cleaner output in case of config errors, avoiding all the noise of a stack trace and instead just saying there's an error and printing the message, which should contain enough information already.

Reviewed By: adamlerer

Differential Revision: D14479641

fbshipit-source-id: 42b1b5a5811132b7eea65571ad70528420de614d