Further changes to the README
Summary:
I asked pemazare to take a look at the previous version of the README and he had very good comments, which I'm addressing here. The "getting started" section is now a walkthrough which describes how to manually do what the fb15k.py script does automatically. It provides the explicit commands, so that users can easily reproduce, and it shows sample outputs and logs, so they can confirm they're doing it correctly.

I then also add a link to the paper, a reference to the doc, paste the BibTex code to cite the paper, remove contractions (it's -> it is), etc.

Reviewed By: adamlerer

Differential Revision: D14542568

fbshipit-source-id: e53bd7e3f039a80b7ea7adfbf56f4dcbd3e46180