Update docs with info we provided in GitHub issues (#40)
Summary:
- [X] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

So that new users can find this information where they expect to find it, and don't have to scour the GitHub issues. Also, this allows to have a page to which to point new questions about this stuff.

Built locally (`make html`), made sure there are no errors, and navigated a bit the result.
Pull Request resolved: https://github.com/facebookresearch/PyTorch-BigGraph/pull/40

Differential Revision: D15088289

Pulled By: lerks

fbshipit-source-id: c9ba9c8db8314d4fc6529c714c58e414db724d89