Move a lot of stuff about batching, training and evaluating to separate modules and classes
Summary: Create a batching module that contains common code for train and eval taking care of splitting and edge list among workers and then into batches. Create Trainer and Evaluator classes to encapsulate the specificities of each, moving a bunch of things out of the model or of the train/eval loops.

Reviewed By: adamlerer

Differential Revision: D15037971

fbshipit-source-id: fa99210a1145d67dbb0435755ef57c664e9a383e