Ease extensibility
Summary:
Allow users to reuse the train loop with customized models and/or training and evaluation code by allowing them to provide any batch processor instance (possibly from a subclass) as trainer and evaluator, and to provide any model instance as model.

Also, make the model's constructor accept any instance of operators and comprarators, so that other subclasses are allowed too, even the ones that are not shipped by default. This basically means moving the config parsing (and thus the enum-to-class translation) out of the constructor.

Reviewed By: adamlerer

Differential Revision: D15122602

fbshipit-source-id: f99d8aba4a809bc9f0479f8a55bdd30131bad13c