Move example scripts into distribution
Summary:
Several users had issues running the FB15k sample script because it is not fully self-contained (it needs other files in the repo) and because it is somewhat coupled to the version of PBG that is installed.

https://github.com/facebookresearch/PyTorch-BigGraph/issues/28
https://github.com/facebookresearch/PyTorch-BigGraph/issues/43

If we just bundle the examples with the main distribution then we solve all of the above.

Reviewed By: adamlerer

Differential Revision: D15152107

fbshipit-source-id: c9322e3c1dd906585b3a118b5d6993be11e9b283