Mitigate race condition around checkpoint version creation
We've experienced occasional failures where, in a distributed setting,
some trainers were seeing the checkpoint version file as empty. Our best
guess is that this is a race condition due to the fact that creating the
file (with an initial empty content) and writing "0" to it are separate
actions and thus other trainer may see the file while it's in a state in
between those two actions.

This commit mitigates the issue in two ways: it avoids creating the file
and writing "0" to it if it doesn't exist (as that's what other trainers
already interpret a missing file as) and if the file is empty it treats
it in the same way as if it were missing.