Mitigate race condition around checkpoint version creation (#58)
Summary:
## Types of changes

- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Motivation and Context / Related issue

We've experienced occasional failures where, in a distributed setting, some trainers were seeing the checkpoint version file as empty. Our best guess is that this is a race condition due to the fact that creating the file (with an initial empty content) and writing "0" to it are separate actions and thus other trainer may see the file while it's in a state in between those two actions.

This commit mitigates the issue in two ways: it avoids creating the file and writing "0" to it if it doesn't exist (as that's what other trainers already interpret a missing file as) and if the file is empty it treats it in the same way as if it were missing.

## How Has This Been Tested (if it applies)

Untested (so far).

## Checklist

- [ ] The documentation is up-to-date with the changes I made.
- [ ] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [ ] All tests passed, and additional code has been covered with new tests.
Pull Request resolved: https://github.com/facebookresearch/PyTorch-BigGraph/pull/58

Differential Revision: D15737307

Pulled By: lerks

fbshipit-source-id: 6be824260f6857c34e14a01fb16d4d6cb916cb13