Make it easier to invoke TSV importer/exporter from user code
The functions that implemented these commands required file paths as
arguments, meaning one could not pass them data that was stored in
memory without serializing it first.

Fixes #50.

I'm also piggybacking a fix to a comment: "m" stood for "metric", but we
are now calling it "comparator", hence use "c".