Start all processes with fork+exec (a.k.a. spawn) rather than just forking
Summary:
When the main train/eval/... function is launched in an environment where there is some other global state or where there are already threads running, forking may cause problems: in the forked subprocesses the state might not match what was expected or could become corrupted (e.g., locks that were held by other threads never being released).

The most resilient solution to this is to perform an exec right after the fork (this is what Python calls the "spawn" method of multiprocessing), so that the global state is fully reinitialized.

The downside is that it takes longer, as the entire interpreter and all the dependencies must be reloaded. In normal operations this would only happen once, at the very beginning, causing a marginal effect. In tests, however, this is more significant. Reducing the number of workers in tests seems to help.

This commit breaks some features (for example the ability to declare custom operators/comparators in the config file) which will be fixed in subsequent commits.

Reviewed By: adamlerer

Differential Revision: D16600947

fbshipit-source-id: c651f5e7466361337160e60c37ee0cd48d3b8dd5