Don't deadlock when worker subprocess dies
Summary:
Add a busy-wait loop that repeatedly checks for the health of the processes in a pool in order to prevent the pool from re-spawning the failing processes indefinitely and becoming stuck.

Also, use the multiprocessing module directly, leaving the torch.multiprocessing import there just so it registers the custom picklers for tensors and storages.

Also, use the proper _classes_ in type annotations, instead of `mp.Process`, `mp.Pool`, etc. which are just functions of the default context (not that this matters much right now: as we don't run a type checker, annotations are basically just for readability).

Reviewed By: adamlerer

Differential Revision: D16600945

fbshipit-source-id: e782178b25ff72b88c846fa768cb1fe676afb01b