Do logging properly
Summary:
Use Python's logging module, rather than plain prints. This helps in ensuring consistency in output location (stdout vs stderr), format, verbosity, etc and allows to reliably re-configure logging behaviours across the board. As a consequence, this enables integration with users' existing logging systems.

To showcase the potential of using a unified standard framework, this commit also adds a component that intercepts all log messages to transparently attach the process name.

As a drawback, since the logging behaviour is now configurable, it needs to be propagated to all subprocesses. Luckily we have the new `subprocess_init` pipes for that, so it's just a matter of plugging it in there.

This commit also finally fixes the oddity of the config parsing function having side-effects (i.e., setting global state).

Reviewed By: adamlerer

Differential Revision: D16687364

fbshipit-source-id: 0015c0dbf8567089caa9f18d39e4f1a4d9e9fe0d