Allow to register custom checkpoint storages
Summary:
Extend to checkpoint storages the same system we have for registering custom operators, comparators, ...

The checkpoint storage will be chosen by the scheme of the path, which is now reinterpreted as an URI (if there's no scheme, then we just assume it's a plain path). Hence a storage for `foo` will handle all URIs of the form `foo://...`.

Reviewed By: adamlerer

Differential Revision: D16961525

fbshipit-source-id: 3ae3110e5b389dffd750296f26f86c4a91346ab9