Fix to actually be able to use non-file edge paths in import_from_tsv
Summary:
In `import_from_tsv` we constructed the output edge paths by converting the paths passed as command-line args (stripping the extension and appending `_partitioned`). Thus, the output edge paths are always file-based. In order to be able to use different schemas we need to be able to separately specify the output edge paths: it makes perfect sense to use the config file for this (we are already using it for the entity path, and may one day use it for the initial data).

The input edge paths (from the command line) and the output ones (from the config) are then matched (i.e., zipped). Thus order now matters. Therefore in the README we explicitly list them, rather than using a glob wildcard.

Reviewed By: adamlerer

Differential Revision: D17502000

fbshipit-source-id: 9cd2e534ab70c600a35e28c4dae57a6015c2bc1e