Make edge importing more efficient
Summary:
As we introduced support for buffered streaming edge appending, we can make use of it to avoid storing all the edges in memory at the same time. This also showcases how to use that streaming API, so that other users can get inspiration from it for their own importers.

Also, increase chunk sizes in HDF5 files from 1MiB to 50MiB. I used 1MiB because it was suggested in the h5py doc but I now suspect that it was a recommendation if one was doing random access to the dataset. This is not our case, so we can use larger chunks, which make writing more efficient.

Reviewed By: adamlerer

Differential Revision: D17525462

fbshipit-source-id: 646e6f733b3262128a17c3ab3372acbe27c8c6a0