Auto-tune num_edge_chunks
Summary: One reason we allow to chunk edgelists is to be able to load edgelists that are too big to fit in memory at once, piece by piece. (There are other reasons however, for example more frequent mixing of edges from different buckets). The former goal can be achieved automatically, that is, given a certain maximum size that edgelists can take in memort, PBG can figure out what is the smallest number of edge chunks that achieve this.

Reviewed By: adamlerer

Differential Revision: D17571778

fbshipit-source-id: e4977078c35f4bdb212ad163acf137eb94d33994