Add optimization to LockServer for bipartite graphs
Summary:
The LockServer used to assign a given partition to one trainer only. Since each trainer typically used two partitions at a time (except for diagonal buckets), this meant one could not have more trainers than half the number of partitions.

For bipartite graphs this is sub-optimal, as when the left- and right-hand side entities are disjoint we can have two trainers each hold one of them without interfering. This then allows to have as many trainers as the number of partitions, doubling the training speed.

As a plus, optimize the code so that it doesn't iterate over all buckets and checks each of them (doing some checks multiple times), but rather only iterate over the buckets whose partitions aren't already locked.

Reviewed By: adamlerer

Differential Revision: D17829706

fbshipit-source-id: c5640900714485ca85c46ea1a81575e305a55667