Use helper to allocate shared memory
Summary: In D18373169 I introduced new shared memory allocations, and we already had quite a few. Each time we do so we need to rely on some hacky internal constructor (i.e., `_new_shared`), and we need to use the old and deprecated `FooStorage` and `FooTensor` classes. Unfortunately this is the only way I know of to allocate shared memory directly. We should however bundle this in a function in order to contain it in a single place, rather than spreading it everywhere. In passing, I also convert the parameter server from `FooTensor`s to dtypes.

Reviewed By: adamlerer

Differential Revision: D18541015

fbshipit-source-id: cb6f52da26f784f95019b51ff6bd01a02a8495f1