Fix schema optional unpacking
Summary:
In some weird circumstances `Optional[type]` is not a singleton class. It only happened in a weird set of multi processing and operators.

Still, I don't think identity equality is guaranteed, so the `!=` fixes that. It's not technically required, but `Union[None, type]` is also a valid way to specify optional, and now this will work properly with that case.

Reviewed By: lerks

Differential Revision: D19799606

fbshipit-source-id: b150219bdc5786fce251aa6036bc0c6ca87231e2