Allow arbitrary dimensions for pbg
Summary:
This is a fairly minimal patch to allow for arbitrary dimensions in pbg

Note it does not translate to filament, i.e. filament still requires everything have the same dimension.

It's possible that I've missed places for verification that things are the correct side, and it's not clear to me how well these will play with gpu filament

Reviewed By: lw

Differential Revision: D20316816

fbshipit-source-id: 77de40b86a68bb36441ecfea73677d5baf6f8d7a