Fix some typing and lint
Summary:
In order to better test my additions, I decided to add pyre checking to pbg.

My guess is that you don't want to roll out pyre to pbg as getting it working consistently on os version and internal would be difficult, and currently would clutter the code with some "ignore this type" nonsense, but this effectively adds changes that help it be type-checked via static analysis.

Before landing, I intend to rebase it off of the pyre config, and clean up any residual ignore comments, but I'm also open to other suggestions.

The changes here can be broken down roughly into a few categories:
1. Enforcing the "NewTypes". These are shims to enforce that an int of one time isn't used when an into of another type is needed. In general these are close to zero cost as it doesn't actually create a new type, but is just a flag for the type checker. That said, it seems like in a recent diff lw changed these to be just type aliases, so these are probably unnecessary.
2. pulling out member variables for refinement. it's common to have code like `if self.field is not None: # something with self.field`, since python allows anything, `self.field` could become None again inside the loop via mutating calls on `self`, so to accurately type check, it must be referenced as a local variable.
3. There were a couple cases of incorrect typing, or typing that became invalid after a refactor
4. A couple instances of changing the name of method parameters in subclasses, which technically changes behavior
5. A couple of named tuples technically override methods / properties of tuples, which invalidates the subclassing, so I changed them to datacalsses. You might not want this as dataclasses were added relatively recently, and it's not clear what version of python you want to support.
6. A few times a type was shadowed, which pyre doesn't handle well
7. A couple times there were complicated control flows where pyre couldn't deduce that a variable would need to be defined.

There was one non trivial error that also wasn't a false positive, which I called out in a comment.

Reviewed By: lw

Differential Revision: D20299182

fbshipit-source-id: b79ddf064ee7e7416cac0cc9cc4a13646ea79787