Make eval_num_XXX_negs optional
Summary:
The validation loss is only correct if eval is performed on the same number of negatives, so provide a way to specify that train and eval negatives should be the same.

For now, I'm not going to take the step of changing the default, because that would be a breaking change.

Reviewed By: lw

Differential Revision: D21262264

fbshipit-source-id: 77b86059b23307a24077f00a9f50988b0124ea91