Allow for single-partition training on 1 GPU machine (simple version)
Summary: GPU training was written to not support "unpartitioned embeddings", but that precludes an important case where you have 1 partition and 1 machine. In that case, you don't have to do any of the special stuff for partitioned (swapping) nor for unpartitioned (sharing). This case is important for e.g. testing with small graphs.

Reviewed By: lw

Differential Revision: D21302580

fbshipit-source-id: 33fccdcc80ff06398e4a07a2fd6d2a654830045e