Use 1 subpart when num_gpus=1
Summary:
The gpu scheduler doesn't handle `num_gpus=1` properly; in that case it should only create 1 subpartition.

(I don't know if we should be subpartitioning automatically based on `num_gpus` in the first place... it's pretty surprising behavior and brittle if the embeddings don't fit in memory with that number of subparts... but that's another story).

lw I'm not super familiar with this part of the code, so please let me know if there's anything I'm missing.

Reviewed By: lw

Differential Revision: D21796318

fbshipit-source-id: ccea39cefa375dec9a5e05fe44d7edf9cde528bd